+++ .github/CHANGELOG.md
@@ -1,9 +1,17 @@
 # Workdrop Changelog
 
-## Unnamed Release
+## Alpha v0.1.0 Release
 
 - initialized Nuxt.js project
 - moved source files to `src/` and added configuration to `nuxt.config.js`
 - installed `node-sass` for Sass styling integration
 - added `dotenv` and configured development server to load setup from `.env`
 - added `mongodb-stitch-browser-sdk`
+- added `mongodb-stitch-browser-services-aws`
+- added `filepond` and `vue-filepond`
+- created file upload handler on `submit` page
+- created request form and request uploader
+- created submission download form and downloader
+- created help, about, privacy policy, and terms pages
+- added icons and meta information
+- pushed for release!